Refactor store inheritance
- `RemoteStore` moved to its own file.
- Create `RemoteFactory` class that `RedisFactory`, `FileFactory`, and
  more can extend to prevent duplicate code.
- `proxystore.store.init_store` can now take type along with enums to
  specify the store.
