+++ proxystore/store/test/test_redis_store.py
@@ -26,7 +26,7 @@ def init() -> None:
 
 def test_redis_store_init() -> None:
     """Test RedisStore Initialization"""
-    RedisStore('redis', REDIS_HOST, REDIS_PORT)
+    RedisStore('redis', hostname=REDIS_HOST, port=REDIS_PORT)
 
     ps.store.init_store(
         ps.store.STORES.REDIS, 'redis', hostname=REDIS_HOST, port=REDIS_PORT
@@ -45,7 +45,7 @@ def test_redis_store_init() -> None:
 
 def test_redis_store_base() -> None:
     """Test RedisStore Base Functionality"""
-    store = RedisStore('redis', REDIS_HOST, REDIS_PORT)
+    store = RedisStore('redis', hostname=REDIS_HOST, port=REDIS_PORT)
     key_fake = 'key_fake'
     value = 'test_value'
 
@@ -87,7 +87,9 @@ def test_redis_store_base() -> None:
 
 def test_redis_store_caching() -> None:
     """Test RedisStore Caching"""
-    store = RedisStore('redis', REDIS_HOST, REDIS_PORT, cache_size=1)
+    store = RedisStore(
+        'redis', hostname=REDIS_HOST, port=REDIS_PORT, cache_size=1
+    )
 
     # Add our test value
     value = 'test_value'
@@ -112,7 +114,9 @@ def test_redis_store_caching() -> None:
     assert store.is_cached(key2)
 
     # Now test cache size 0
-    store = RedisStore('redis', REDIS_HOST, REDIS_PORT, cache_size=0)
+    store = RedisStore(
+        'redis', hostname=REDIS_HOST, port=REDIS_PORT, cache_size=0
+    )
     store.set(value, key=key1)
     assert store.get(key1) == value
     assert not store.is_cached(key1)
@@ -174,10 +178,17 @@ def test_redis_factory() -> None:
 
     # Clear store to see if factory can reinitialize it
     ps.store._stores = {}
-    f = RedisFactory(key, 'redis', REDIS_HOST, REDIS_PORT)
+    f = RedisFactory(
+        key, 'redis', store_kwargs={'hostname': REDIS_HOST, 'port': REDIS_PORT}
+    )
     assert f() == [1, 2, 3]
 
-    f2 = RedisFactory(key, 'redis', REDIS_HOST, REDIS_PORT, evict=True)
+    f2 = RedisFactory(
+        key,
+        'redis',
+        store_kwargs={'hostname': REDIS_HOST, 'port': REDIS_PORT},
+        evict=True
+    )
     assert store.exists(key)
     assert f2() == [1, 2, 3]
     assert not store.exists(key)
@@ -185,7 +196,9 @@ def test_redis_factory() -> None:
     store.set([1, 2, 3], key=key)
     # Clear store to see if factory can reinitialize it
     ps.store._stores = {}
-    f = RedisFactory(key, 'redis', REDIS_HOST, REDIS_PORT)
+    f = RedisFactory(
+        key, 'redis', store_kwargs={'hostname': REDIS_HOST, 'port': REDIS_PORT}
+    )
     f.resolve_async()
     assert f._obj_future is not None
     assert f() == [1, 2, 3]
